+++ src/base/operation.js
@@ -31,6 +31,8 @@ import { ManageCreatePollRequestBuilder } from './operations/manage_create_poll_
 import { ManageVoteBuilder } from './operations/manage_vote_builder'
 import { ManagePollBuilder } from './operations/manage_poll_builder'
 import { ManageAccountSpecificRuleBuilder } from './operations/manage_account_specific_rule_builder'
+import { InitiateKYCRecoveryBuilder } from './operations/initiate_kyc_recovery_builder'
+import { CreateKYCRecoveryRequestBuilder } from './operations/create_kyc_recovery_request_builder'
 
 export class Operation extends BaseOperation {
   /**
@@ -454,6 +456,12 @@ export class Operation extends BaseOperation {
         break
       case xdr.OperationType.manageAccountSpecificRule():
         ManageAccountSpecificRuleBuilder.manageAccountSpecificRuleToObject(result, attrs)
+        break
+      case xdr.OperationType.initiateKycRecovery():
+        InitiateKYCRecoveryBuilder.initiateKYCRecoveryToObject(result, attrs)
+        break
+      case xdr.OperationType.createKycRecoveryRequest():
+        CreateKYCRecoveryRequestBuilder.createKYCRecoveryRequestOpToObject(result, attrs)
         break
       default:
         throw new Error('Unknown operation')