+++ src/const/key-value.const.js
@@ -8,5 +8,8 @@ export const KEY_VALUE_KEYS = Object.freeze({
   atomicSwapBidTasks: 'atomic_swap_bid_tasks:*',
   atomicSwapAskTasks: 'atomic_swap_ask_tasks',
   paymentCreateTasks: 'payment_create_tasks:*',
-  createPollTasks: 'create_poll_tasks:*'
+  createPollTasks: 'create_poll_tasks:*',
+  kycRecoveryEnabled: 'kyc_recovery_enabled',
+  kycRecoverySignerRole: 'kyc_recovery_signer_role',
+  createKycRecoveryTasks: 'create_kyc_recovery_tasks'
 })